requirements.txt
streamlit
openai-whisper
moviepy
ffmpeg-python
srt
openai
